shellIntegration-bash.sh: fix incorrect check for __vsc_preexec_only trap
The function `__vsc_preexec` is never set as a trap, but
`__vsc_preexec_only` and `__vsc_preexec_all` are. In fact,
`__vsc_preexec` is *only* referenced by those trap functions, and in no
other context.

If `__vsc_preexec_only` were set as the existing trap, the previous
condition would cause it to be called twice. (Which is harmless, but
unintended).