shellIntegration-bash.sh: avoid clobbering existing __vsc_dbg_trap
The condition checking whether the existing trap was `__vsc_preexec_all`
was performed only *after* replacing the `$__vsc_dbg_trap` variable.
That variable is eval'd by that function and exists to wrap/preserve the
original trap expression; and in that case would be clobbered by a
useless self-recursive call.

The condition only gated an ineffective 'are we setting this trap to its
current value' condition, rather than the clearly-intended 'would we be
clobbering an already-wrapped trap' condition.

This avoids that situation by avoiding updating the `__vsc_dbg_trap`
variable unless it's different from the trap we're about to set.